Create ex 1 and 2
My name is Kirsty, so the first and last letter of my name (K + Y = 11 + 25) give the number  36. The 36th Fibonacci number is 14930352.
My surname is O'Brien
The first letter O is number 79
The last letter n is number 110
Fibonacci number 189 is 1409869790947669143312035591975596518914

ord() = Given a string representing one Unicode character, return an integer representing the Unicode code point of that character.